Refactor naming and folder for RuleDesigner => ModifyExport
